Revert empty() check to isset() for cmd
We still need to be able to pass an empty string to the terminal. It would be much more hack-n-slash to prevent sending a blank command, and I personally feel it is more like a real terminal when you are able to send a blank command. If empty() is used, the code would also need to be modified further to prevent it from returning the full console HTML, as it would fail the php check and never hit die() if a blank command is sent.